Do not throw fatal error if key is unused (#3)
In the following php code:
```php
$Language->getText('plugin_git', 'Unable to update template');
```

`Unable to update template` is not a valid secondary key. It should
be `view_admin_template_cannot_update` according to the .tab file:
```
plugin_git	view_admin_template_cannot_update	Unable to update template
```

In this case the tab2gettext should not end up in a cryptic fatal error.